[systemd] Install an empty machine-id. Contribute to JB#38016
From machine-id(5):

For operating system images which are created once and used on multiple
machines, for example for containers or in the cloud, /etc/machine-id
should be an empty file in the generic file system image. An ID will be
generated during boot and saved to this file if possible. Having an
empty file in place is useful because it allows a temporary file to be
bind-mounted over the real file, in case the image is used read-only.